fixes for unintended errors before shutdown of lifecycle node taken from previous PRs.

+++ nav2_map_server/include/nav2_map_server/map_3d/map_io_3d.hpp
@@ -34,8 +34,8 @@ namespace map_3d
 
 struct Origin
 {
-  std::vector<float> center;
-  std::vector<float> orientation;
+  std::vector<double> center;
+  std::vector<double> orientation;
 
   void resize()
   {
